Icon List - Class migration
