+++ src/blocks/icon-list/attributes.js
@@ -32,6 +32,10 @@ const attributes = {
 	block_id: {
 		type: "string"
 	},
+	classMigrate: {
+		type: "boolean",
+		default: false
+	},
 	align: {
 		type: "string",
 		default: "left"
